Got the service to start and stop based on if the app is running.  This is to prevent any possible attack on the phone while the user walks around with the app running in the background.
Now, here's the issue; The messages are received, processed and responded to from within a HostApduService (which runs in it's own intent).  Seemingly, the only time that you can send a message between intents is during the spawn/starting of that service.  So the issue remains; how do we pass the information that each establishment would need to transmit for identification/authorization and response from the main activity?  I have a few ideas, but they are all hacks (And I'd like to stay away from writing to the disk).  There must be a better way to do it.